Moved the Game class to the elements package
